pyproject.toml
[project]
name = "zania-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.4",
    "langchain>=0.3.7",
    "langchain-community>=0.3.5",
    "faiss-cpu>=1.9.0",
    "pypdf>=5.1.0",
    "aiofiles>=24.1.0",
    "fastembed>=0.4.1",
    "loguru>=0.7.2",
    "jq>=1.8.0",
    "langchain-openai>=0.2.6",
    "langchain-text-splitters>=0.3.2",
    "python-dotenv>=1.0.1",
]
